refactor: Phase 3 - Move keyboard UI components to ui/keyboard/
- Move TibetanKeyboardView, AIKeyboardView, AIKeyboardCodes to ui/keyboard/
- Move EmojiKeyboardView, EmojiItemList, SimpleSearchKeyboard to ui/keyboard/
- Move CorrectionsAdapter, KeyboardType, AIKeyboardInterface to ui/keyboard/
- Update all package declarations and imports
