+++ app/src/main/java/com/kharagedition/tibetankeyboard/ui/keyboard/EmojiKeyboardView.kt
@@ -1,4 +1,6 @@
-package com.kharagedition.tibetankeyboard
+package com.kharagedition.tibetankeyboard.ui.keyboard
+
+import com.kharagedition.tibetankeyboard.R
 
 import android.content.Context
 import android.graphics.BitmapFactory
